requirements.txt
fastapi
uvicorn
httpx
python-dotenv
openai
httpx
fpdf2
jinja2
python-multipart